requirements.txt
ImageHash==4.0
Pillow==5.2.0
pybktree==1.1
sanic==0.8.1